MS-129: rename new version of shared components
