+++ src/Components/Admin/Positions/Form/index.js
@@ -2,10 +2,10 @@ import { useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import useQuery from 'Hooks/useQuery';
 import styles from './form.module.css';
-import Input from 'Components/Shared/InputV2';
+import Input from 'Components/Shared/Input2';
 import Button from 'Components/Shared/Button';
 import Modal from 'Components/Shared/Modal';
-import Select from 'Components/Shared/SelectV2';
+import Select from 'Components/Shared/Select2';
 import { useDispatch, useSelector } from 'react-redux';
 import { Form, Field } from 'react-final-form';
 import { getPositionById, createPosition, updatePosition } from 'redux/positions/thunks';
